+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [x.y.z] - yyyy-mm-dd
 ### Added
 * Added GitHub Publish Workflow
+* Added Primer Figures 3.10a and 3.10b and relative tests
 
 ### Changed
 * Changed topological_sort as method of DirectedGraph class